# docs/assets/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/training/projects/{projectId}/validate": {
      "get": {
        "operationId": "TrainingController_validateProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Labeling project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result indicating whether the project is ready for training",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Validate project training data",
        "tags": [
          "Training"
        ]
      }
    },
    "/api/training/projects/{projectId}/train": {
      "post": {
        "operationId": "TrainingController_startTraining",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Labeling project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTrainingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training job created and started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Start model training for a project",
        "tags": [
          "Training"
        ]
      }
    },
    "/api/training/projects/{projectId}/jobs": {
      "get": {
        "operationId": "TrainingController_getTrainingJobs",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Labeling project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of training jobs for the project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingJobDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get all training jobs for a project",
        "tags": [
          "Training"
        ]
      }
    },
    "/api/training/jobs/{jobId}": {
      "get": {
        "operationId": "TrainingController_getJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Training job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Training job details and current status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get training job status",
        "tags": [
          "Training"
        ]
      },
      "delete": {
        "operationId": "TrainingController_cancelJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Training job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Training job cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelJobResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Cancel a training job",
        "tags": [
          "Training"
        ]
      }
    },
    "/api/training/projects/{projectId}/models": {
      "get": {
        "operationId": "TrainingController_getTrainedModels",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Labeling project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of trained models produced from this project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainedModelDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get trained models for a project",
        "tags": [
          "Training"
        ]
      }
    },
...
  },
  "info": {
    "title": "Document Intelligence Platform API",
    "description": "Comprehensive REST API for document processing, OCR, workflow orchestration, model training, and human-in-the-loop review",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "keycloak-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ValidationResultDto": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Whether the project has sufficient labeled data for training"
          },
          "labeledDocumentsCount": {
            "type": "number",
            "description": "Number of labeled documents in the project"
          },
          "minimumRequired": {
            "type": "number",
            "description": "Minimum number of labeled documents required"
          },
          "issues": {
            "description": "List of validation issues preventing training",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "valid",
          "labeledDocumentsCount",
          "minimumRequired",
          "issues"
        ]
      },
      "StartTrainingDto": {
        "type": "object",
        "properties": {}
      },
      "TrainingJobDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Training job ID"
          },
          "projectId": {
            "type": "string",
            "description": "Labeling project ID"
          },
          "status": {
            "type": "string",
            "description": "Training job status",
            "enum": [
              "PENDING",
              "UPLOADING",
              "UPLOADED",
              "TRAINING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "containerName": {
            "type": "string",
            "description": "Azure blob storage container name"
          },
          "sasUrl": {
            "type": "string",
            "description": "SAS URL for blob storage access"
          },
          "blobCount": {
            "type": "number",
            "description": "Number of blobs uploaded for training"
          },
          "modelId": {
            "type": "string",
            "description": "Azure model ID once training completes"
          },
          "operationId": {
            "type": "string",
            "description": "Azure training operation ID"
          },
          "errorMessage": {
            "type": "string",
            "description": "Error message if training failed"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When training started"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When training completed"
          }
        },
        "required": [
          "id",
          "projectId",
          "status",
          "containerName",
          "blobCount",
          "startedAt"
        ]
      },
      "TrainedModelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Trained model record ID"
          },
          "projectId": {
            "type": "string",
            "description": "Labeling project ID"
          },
          "trainingJobId": {
            "type": "string",
            "description": "Training job that produced this model"
          },
          "modelId": {
            "type": "string",
            "description": "Azure Document Intelligence model ID"
          },
          "description": {
            "type": "string",
            "description": "Model description"
          },
          "docTypes": {
            "type": "object",
            "description": "Azure document type definitions",
            "additionalProperties": true
          },
          "fieldCount": {
            "type": "number",
            "description": "Number of fields the model recognizes"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the model was created"
          }
        },
        "required": [
          "id",
          "projectId",
          "trainingJobId",
          "modelId",
          "fieldCount",
          "createdAt"
        ]
      },
      "CancelJobResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}